requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
Pillow
numpy
torch
python-dotenv
